Explicitly explain switching on enums and tagged structs
Previously, the first place we encountered switching on enums was in the context of tagged structs, where there were two layers of complexity involved. Instead, explain how switching on enums works in the enums section, and then explain how captures layer onto that in the tagged structs section, explicitly pointing out its similarity to switching on the enum.