Chapter 1: Clarify some language around payload captures (#111)
* Clarify some language around payload captures

Payload captures were mentioned a couple times before actually being explained, without actually explaining what they were. The first time they're mentioned, this commit adds a brief explanation of what bit of code is the payload capture and mentions we'll go more in depth later, and elaborates a bit more on what we're doing the second time they come up.

It also rewrites the section on copy semantics and pointer captures to instead focus on captures values being immutable by default.

* Be a little clearer around dereferencing

My original language here was a little misleading I think

* Fix apostrophe

* Explain pointer captures better

Call out that we're using a different form of payload capture the first time we use pointer capture, and remove another reference in the optionals section to payload captures being copied (they are not), replacing it with another pointer capture to demonstrate that we can modify the value in the optional via a capture.

* Simplify explanation of pointer capture below

We've already explained it above, so just reference that here, and take the chance to explain more explicitly that the pointer is still immutable but the value referenced is not.