Update 11-pointers.mdx
fix typo in code snippet. Discarding the variable should be `_ = &y;` instead of `_ = y;`.

replace `var x: u16 = 0;` with `var x: u16 = 5; x -= 5;`. Else the compiler complains with `error: local variable is never mutated`. The updated example also highlights that the compiler is able to detect the value of x is not 0 through a mutation